requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attr==0.3.2
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
DateTime==5.5
emoji==2.14.0
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
google==3.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
munch==4.0.0
pillow==11.1.0
propcache==0.2.1
psutil==6.1.1
py-cpuinfo==9.0.0
pycryptodome==3.21.0
Pygments==2.19.1
python-dotenv==1.0.1
pytz==2024.2
rel==0.4.9.19
requests==2.32.3
setuptools==75.7.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.1
yarl==1.18.3
youtube-dl==2021.12.17
youtube-search==2.1.2
youtube-search-python==1.6.6
yt-dlp==2024.12.23
ytmusicapi==1.9.1
zlapi==1.0.3
zope.interface==7.2
